refactor: #82 구조 변경
