+++ src/main/java/org/team1/nbe1_2_team01/global/auth/email/repository/EmailRepository.java
@@ -1,10 +1,9 @@
-package org.team1.nbe1_2_team01.global.auth.redis.repository;
+package org.team1.nbe1_2_team01.global.auth.email.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import org.team1.nbe1_2_team01.global.auth.redis.token.EmailToken;
+import org.team1.nbe1_2_team01.global.auth.email.token.EmailToken;
 
 import java.util.Optional;
-import java.util.UUID;
 
 public interface EmailRepository extends CrudRepository<EmailToken,String> {
     boolean existsByCode(String code);
