+++ src/test/java/org/team1/nbe1_2_team01/domain/user/controller/UserControllerTest.java
@@ -24,9 +24,8 @@ import org.team1.nbe1_2_team01.domain.user.repository.CourseRepository;
 import org.team1.nbe1_2_team01.domain.user.repository.UserRepository;
 import org.team1.nbe1_2_team01.domain.user.service.UserService;
 import org.team1.nbe1_2_team01.global.auth.jwt.service.JwtService;
-import org.team1.nbe1_2_team01.global.auth.login.service.LoginService;
-import org.team1.nbe1_2_team01.global.auth.redis.repository.RefreshTokenRepository;
-import org.team1.nbe1_2_team01.global.auth.redis.token.RefreshToken;
+import org.team1.nbe1_2_team01.global.auth.jwt.respository.RefreshTokenRepository;
+import org.team1.nbe1_2_team01.global.auth.jwt.token.RefreshToken;
 import java.util.Optional;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
